// src/app/models/Employee.ts
export class Employee {
  empId: number;
  name: string;
  city: string;
  state: string;
  email: string;
  contact: string;
  address: string;
  pinCode: string;
  constructor() {
    this.empId = 1;
    this.name = '';
    this.email = '';
    this.contact = '';
    this.address = '';
    this.city = '';
    this.state = '';
    this.pinCode = '';
  }
}